+++ src/tools/teams.ts
@@ -337,7 +337,7 @@ export class TeamsTools {
       };
     }
 
-    await this.repo.deleteChannelAsync(result.token!.targetId);
+    await this.repo.deleteChannelAsync((result.token!.targetId as number));
     return {
       content: [{
         type: 'text' as const,
@@ -441,7 +441,7 @@ export class TeamsTools {
       };
     }
     const { body, contentType } = result.token!.metadata as { body: string; contentType: string };
-    const messageId = await this.repo.sendChannelMessageAsync(result.token!.targetId, body, contentType);
+    const messageId = await this.repo.sendChannelMessageAsync((result.token!.targetId as number), body, contentType);
     return {
       content: [{
         type: 'text' as const,
@@ -505,7 +505,7 @@ export class TeamsTools {
       };
     }
     const { body, contentType } = result.token!.metadata as { body: string; contentType: string };
-    const replyId = await this.repo.replyToChannelMessageAsync(result.token!.targetId, body, contentType);
+    const replyId = await this.repo.replyToChannelMessageAsync((result.token!.targetId as number), body, contentType);
     return {
       content: [{
         type: 'text' as const,
@@ -609,7 +609,7 @@ export class TeamsTools {
       };
     }
     const { body, contentType } = result.token!.metadata as { body: string; contentType: string };
-    const messageId = await this.repo.sendChatMessageAsync(result.token!.targetId, body, contentType);
+    const messageId = await this.repo.sendChatMessageAsync((result.token!.targetId as number), body, contentType);
     return {
       content: [{
         type: 'text' as const,
@@ -700,7 +700,7 @@ export class TeamsTools {
       };
     }
     const { reaction_type, message_type } = result.token!.metadata as { reaction_type: string; message_type: 'channel' | 'chat' };
-    await this.repo.addMessageReactionAsync(result.token!.targetId, message_type, reaction_type);
+    await this.repo.addMessageReactionAsync((result.token!.targetId as number), message_type, reaction_type);
     return {
       content: [{
         type: 'text' as const,